Made some stuff nicer. Removed old stuff that wasn't ever used to make startup a bit lighter.
