+++ src/main/java/net/kodehawa/mantarobot/commands/info/stats/StatsHelper.java
@@ -14,7 +14,7 @@
  * along with Mantaro.  If not, see http://www.gnu.org/licenses/
  */
 
-package net.kodehawa.mantarobot.commands.info;
+package net.kodehawa.mantarobot.commands.info.stats;
 
 import net.dv8tion.jda.core.entities.MessageEmbed;
 import net.dv8tion.jda.core.events.message.guild.GuildMessageReceivedEvent;
