+++ src/main/java/net/kodehawa/mantarobot/core/listeners/MantaroListener.java
@@ -41,8 +41,8 @@ import net.dv8tion.jda.core.hooks.EventListener;
 import net.kodehawa.mantarobot.MantaroBot;
 import net.kodehawa.mantarobot.commands.currency.RateLimiter;
 import net.kodehawa.mantarobot.commands.custom.EmbedJSON;
-import net.kodehawa.mantarobot.commands.info.GuildStatsManager;
-import net.kodehawa.mantarobot.commands.info.GuildStatsManager.LoggedEvent;
+import net.kodehawa.mantarobot.commands.info.stats.manager.GuildStatsManager;
+import net.kodehawa.mantarobot.commands.info.stats.manager.GuildStatsManager.LoggedEvent;
 import net.kodehawa.mantarobot.core.MantaroCore;
 import net.kodehawa.mantarobot.core.listeners.command.CommandListener;
 import net.kodehawa.mantarobot.core.listeners.entities.CachedMessage;
